Fixing sample
