+++ dart-sample/app/src/main/AndroidManifest.xml
@@ -5,8 +5,6 @@
     android:versionName="1.0"
     package="com.f2prateek.dart.example">
 
-  <uses-sdk android:minSdkVersion="11" android:targetSdkVersion="19"/>
-
   <application
       android:label="@string/app_name"
       android:name=".SampleApp">
