+++ dart-sample/module1-navigation/src/main/java/com/f2prateek/dart/module1/Module1ActivityNavigationModel.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.f2prateek.dart;
+package com.f2prateek.dart.module1;
 
 import android.support.annotation.Nullable;
 import android.util.SparseArray;
