+++ dart-sample/module1-navigation/src/main/java/com/f2prateek/dart/module1/Module1ServiceNavigationModel.java
@@ -15,21 +15,11 @@
  * limitations under the License.
  */
 
-package com.f2prateek.dart;
+package com.f2prateek.dart.module1;
 
-import android.app.IntentService;
-import android.content.Intent;
-import dart.Dart;
+import dart.BindExtra;
 
-public class Module1Service extends IntentService {
+public class Module1ServiceNavigationModel {
 
-  private Module1ServiceNavigationModel navigationModel = new Module1ServiceNavigationModel();
-
-  public Module1Service() {
-    super("Module1Service");
-  }
-
-  @Override protected void onHandleIntent(Intent intent) {
-    Dart.bindNavigationModel(navigationModel, intent.getExtras());
-  }
+  @BindExtra String stringExtra;
 }
